WIP: property mangler
I've started to implement the property mangler and thought it'd be nice to open a PR.

I'll probably need to start using JsWord instead of regular strings, but sometimes we're mangling strings as well.

I started with keep_quoted: 'strict' which is the most sane option, and also the simplest. I wonder if we need to have all the others as well.