make test scripts faster when the runner wants to filter tests
When developing the property mangler, I'd commented out the lines in this 'if' statement so I could run tests faster.

So I came up with this PR to enable us to filter tests when desired, and run them quickly, but also allow to run everything by omitting filter arguments.

If you have a better idea for testing quickly please let me know :) this is still pretty slow.