Use package buckets for browser cache info.
And test that requests are within the same package before processing
rewrites and ignores otherwise we could inadvertently use a rewrite or
ignore in a package outside of the one declaring the `browser` object.

Also disable the recursive resolve() for handling keys that are module
references as it can cause a segfault in some situations.