fix(es/lexer): Fix lexing of e numbers with large exponent (#3060)
Exponents so large that they become INFINITY when parsed as
Rust's f64 made lexer panic. The reason is that it parses an
exponent as f64, then converts it back to string to make a string
representation of the number being parsed, then parses the
resulting string as f64 again. INFINITY became 'inf' when
converted to string, so, after appending that to the resulting
string it became something like '1e+inf'. This is invalid and
parsing it as f64 panicked.
Fix that by first checking if exponent became INFINITY after
parsing and, it that case, just returning a fixed value instead of
trying to form the string and parse it.
Fixed value is determined as follows:
 * it is INFINITY if exponent was positive and mantissa
   was nonzero
 * it is 0.0 otherwise (is exponent was negative or mantissa was
   zero)