compress: Fix unsound optimization of `new RegExp(…)`
In #4932, an optimization was added that transforms `new RegExp(…)` to
`RegExp(…)`.  It seems to have been motivated by this sentence in the
ECMAScript specification:

“Thus the function call RegExp(…) is equivalent to the object creation
expression new RegExp(…) with the same arguments.”

But that sentence turns out to be wrong.  It contradicts the detailed
steps in the following section of the specification, as well as the
empirical behavior of JavaScriptCore, SpiderMonkey, and V8:

    > r = /a/;
    /a/
    > RegExp(r) === r
    true
    > new RegExp(r) === r
    false

Restrict this optimization to the case where we can prove one of the
first two arguments is a string.

Fixes #6941.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>