fix(plugin/runner): Disable wasi env cleanup (#7607)
**Description:**

Even after setting shared runtime for the wasi env, `cleanup` still panics. I suspect something internal to cleanup doesn't honor those handles correctly; for now, disable it and try to see if this works.

As commented, this is not a huge issue as the plugin usually doesn't have resources to be cleaned up anyway.