fix(es/helpers): Add `addInitializer` to field decorators (#8619)
**Description:**

At some point the spec text of ES decorators was changed to also have
addInitializer for field decorators. This is already used e.g. in mobx.
This simple change removes the special case that was present for field
decorators and seems to fix the issue for me.

This change is backwards compatible since it only affects code that was
broken before.

**Related issue:**

 - Closes #8525