chore(es/codegen): Bump minimum required swc_allocator version to 0.1.8 (#9492)
**Description:**

`swc_ecma_codegen` uses `swc_allocator::collections` modules:


https://github.com/swc-project/swc/blob/f500dd4530a1e214d787151689ef9ddb64fb7510/crates/swc_ecma_codegen/src/text_writer/basic_impl.rs#L3


https://github.com/swc-project/swc/blob/f500dd4530a1e214d787151689ef9ddb64fb7510/crates/swc_ecma_codegen/tests/sourcemap.rs#L5

`collections` module was added to `swc_allocator` in 0.1.8. However,
`swc_ecma_codegen` still says that the minimum required version of
`swc_allocator` is 0.1.7:


https://github.com/swc-project/swc/blob/f500dd4530a1e214d787151689ef9ddb64fb7510/crates/swc_ecma_codegen/Cargo.toml#L27

This can cause a compile error when someone has a lockfile that pins
`swc_allocator` 0.1.7, which is supposed to work with `swc_ecma_codegen`
according to the manifest, but actually not.