fix(es/minifier): Fix minification of `framer-motion` by checking `cons.termniates()` (#9818)
**Description:**

Next.js 15 seems to minify parts of `framer-motion`'s code in an invalid
way, changing the meaning. The original function modifies the single
`definition` variable as it changes meaning (see `input.js`):
- if `definition` is a function, it calls it and saves the return value
as `definition`
- if `definition` is a string, it resolves animation variant based on
value as name and saves the variant as `definition`
- if `definition` is again a function, it calls it again and saves the
return value as `definition` again

However, `swc` seems to be not noticing that `definition` may be mutated
between the calls and collapses all `if`s into one:

```
// `resolveVariantFromProps` in framer-motion
// https://github.com/motiondivision/motion/blob/ecd97f7dce8954be300aa73ab6a96208437941c5/packages/framer-motion/src/render/utils/resolve-variants.ts#L31-L72
export function resolveVariantFromProps(props, definition, custom, visualElement) {
    if ("function" == typeof definition || ("string" == typeof definition && (definition = props.variants && props.variants[definition]), "function" == typeof definition)) {
        const [current, velocity] = getValueState(visualElement);
        definition = definition(void 0 !== custom ? custom : props.custom, current, velocity);
    }
    return definition;
}
```

which skips the first call to `definition`, first "resolution" step.

Note: I suspect the `output.js` may not be precisely what SWC may output
as minified code (maybe there is a way to minify it further). It
definitely differs from the current output though by the initial call
and reassignment to `definition`.

---------

Co-authored-by: austaras <austaras@outlook.com>