perf(es/minifier): Make analyzer not call `collect_infects_from` recursively (#9924)
**Description:**

We don't need to make time complexity non-linear by recursively visiting the effects of infection. Instead, suppose we stop the infection analyzer when the usage analyzer invokes the infection analyzer. In that case, we will have practically identical analysis results because the points are named using an id, and it points to all the infections that occurred by referencing the id.

**Related issue:**
 - Closes https://github.com/swc-project/swc/issues/9927