fix(es/loader): Fix the absolute path check when resolving modules (#10080)
fixes: https://github.com/swc-project/swc/issues/9854

`if !module_specifier.starts_with('.')` is true when `module_specifier`
is neither a relative path nor an absolute path like `classnames`, then
the file in current working directory will be found.

I gave up adding a test because it needs to change the current working
directory (by calling `std::env::set_current_dir`), which causes
failures of other tests.

Another changed line of code is to align with
https://www.typescriptlang.org/docs/handbook/modules/reference.html#baseurl,
which says:
> The baseUrl compiler option can be combined with any moduleResolution
mode and specifies a directory that bare specifiers (module specifiers
that don’t begin with **./, ../, or /**) are resolved from