fix(es/preset-env): Consider `browserslist` config if `env.target` is not configured (#8921)
**Description:**

The `env.path` option in `.swcrc` was originally implemented for `browserslist`, but implementation was (unintentionally?) removed during the migration to `browserslist-rs` (#2845).
Searching the current directory for browserslist configuration was also dropped in the process.

This PR reimplements searching for browserslist configuration using the lookup functionality provided by `browserslist-rs`


**Things to consider:**
- only works for non-wasm targets as `browserslist-rs` does not provide configuration loading for wasm32 targets (which is probably to be expected that we won't check environment variables or search a filesystem in that case)
- `feat(preset-env): Let browserslist-rs handle default config path` changes the default for an unconfigured `env.path` from an empty string for wasm targets / current dir for other targets to a `None` value.
`browserslist-rs` already looks up the current dir if `path` is `None` and it felt better to let it handle this case itself.

The commit which migrated from `browserslist` to `browserslist-rs` mentioned fixing the `default_path`, but does not mention what was broken, and I also did not find any other usage of the `path` property. If the previous implementation was intentional, ignore that last commit in the PR, the first two sufficiently address the bug.

**Related issue:**

- Closes https://github.com/swc-project/swc/issues/3365

---------

Co-authored-by: Donny/강동윤 <kdy1997.dev@gmail.com>