Fix sourcemap resolution for errors thrown from constructor
Previously, if an error was thrown from the constructor, the sourcemap
callstack would show the error coming from `at new constructor`.

This fix allows the error to be more specific and instead refer to the
class name. i.e. if a class `ReactComponent` throws an error in
`constructor`, the callstack will now say `at new ReactComponent`.