fix(es/codegen): Exclude `constructor` from source map names (#10551)
**Description:**

Previously, if an error was thrown from the constructor, the sourcemap
callstack would show the error coming from `at new constructor`.

This fix allows the error to be more specific and instead refer to the
class name. i.e. if a class `ReactComponent` throws an error in
`constructor`, the callstack will now say `at new ReactComponent`.

```
~/sourcemapped-stackframes-playground/src/class-component.js:3
    throw new Error("Boom");
          ^
Error: Boom
    at new ReactComponent (/Users/joshuahannaford/work/sourcemapped-stackframes-playground/src/class-component.js:3:11)
    at new Component (~/work/sourcemapped-stackframes-playground/src/class-component.js:7:1)
    at Object.<anonymous> (~/work/sourcemapped-stackframes-playground/src/class-component.js:9:1)
    at Module._compile (node:internal/modules/cjs/loader:1554:14)
```


**Related issue:**

 - Closes https://github.com/swc-project/swc/issues/9748

This partially addresses the user report in this error. It fixes
`constructor` being used as the method name rather than the class name
as it should be. What this doesn't fix though is the mangled names being
used instead of the original source names.

With the settings that user is using, neither `terser`, `esbuild`, nor
`babel` create sourcemaps that properly map between the original code
and the mangled names. In order to get the expected behaviour, they
would need to modify their configuration to not mangle top-level
elements using `topLevel: false`. This PR makes our output consistent
with other build systems with this configuration setup.

---------

Co-authored-by: Donny/강동윤 <kdy1997.dev@gmail.com>