fix(ts/isolated-dts): Add edges `SymbolFlags::Value` and `SymbolFlags::Type` in exports (#10577)
**Description:**

We should add edges for all three possible namespaces because there's no mechanism to merge SymbolFlags with the same IDs.

For example: 
```ts
const a = 1 // SymbolFlags::Value
type a = number // SymbolFlags::Type
class A {} // SymbolFlags::all()

export { a, A }
```

Ideally, we should merge two `a`s with `SymbolFlags::all`. But it's not efficient to merge nodes in the real time. So when exporting symbols, we add edges for all possible cases.

**Related issue:**

 - Closes https://github.com/swc-project/swc/issues/10576