Mark `atom!` allocated Atoms as having a 'static' lifetime and stop refcounting them
The Clone and Drop routines will treat them like `inline` atoms but their storage will instead be in a `static` item.

This should speed up initialization and access to `atom!` values since there is no more lazy lock just a simple pointer tag.  This does complicate `eq` and `hash` a bit as well as `as_str` to handle the new location

This follows a similar pattern used in turbopack as of https://github.com/vercel/next.js/pull/81994.  Things are different here due to the use of `ThinArc` for storage, which required me to introduce a separate struct for the `&'static str` case.