perf(es/parser): Introduce checkpoint to reduce clone (#11001)
**Description:**

There could be no need to clone entire lexer and parser when calling
`try_parse_xxx` in typescript. Here introduces checkpoints which only
contains necessary data to recover from failed trials. This is also the
prerequisite for https://github.com/swc-project/swc/pull/11000. And we
can also remove the `Rc<RefCell<T>>` in the lexer later

I'm not sure the data is completed for correctness. Let's see the test
results.