test(es): Make `isModule` in test cases available (#11038)
**Description:**

I think this is a mistake when I'm fixing the ci in
https://github.com/swc-project/swc/pull/11035.

1. Config pass directly with rust api will merge the config read from
`.swcrc` in `build_as_input`
https://github.com/swc-project/swc/blob/0f177df7f3fc02df8895eaadd1ebe8f1f066985e/crates/swc/src/config/mod.rs#L276
2. From the `Merge` impl for `IsModule`, `IsModule::Unknown` has higher
priority than `IsModule::Bool(true)`
https://github.com/swc-project/swc/blob/0f177df7f3fc02df8895eaadd1ebe8f1f066985e/crates/swc_config/src/is_module.rs#L23
3. `IsModule::Unknown` is passed to this test code, which means all
`isModule` in the `.swcrc` of each test case will not take effect.

I think the origin intention is that for these test cases, if there's no
`isModule` in `.swcrc`, then `IsModule::Unknown` is used by default.
However, it's actually hard to do so. As is said above, if we don't
provide `IsModule` the default value is used, and if we provide
`IsModule::Unknown` the `.swcrc` will become invalid.

Finally I just revoke `IsModule`. I hope that it's suitable to assume
each test case is a module...