fix(hstr): Skip only `\u` for unicode (#11216)
**Description:**

Attempt to fix #11214.

The reason for unespaced `\\` was special handling for unicode `\u`
codepoint, that escapes any string literal starts with `\u` - for
windows, path separator with u (`\\u`) matches with this case.

PR attempts to solve by lookahead, confirming all 4 following char is
hex digit to represent unicode hex, otherwise consider it as plain
string. To be honest I'm not sure if this is acceptable approach or not,
feel free to close if there's better way to fix.

**Related issue:**

 - Closes #11214