fix(es/parser): Don't call `bump_bytes` in the `continue_if` of `byte_search!` (#11328)
**Description:**

In https://github.com/swc-project/swc/pull/11323, we only record the
`pos` as an offset related to the current remaining string and bumps the
lexer at the end of `byte_search!` to reduce the calls to `bump_bytes`.
However, if `bump_bytes` is manually called in the `continue_if`, the
invariant is broken.